+++ commands/image_processing/gay.js
@@ -1,19 +1,21 @@
-const Discord = require("discord.js");
+const Discord = module.require("discord.js");
 const fs = require("fs");
 
-module.exports.run = async (client, message, args, prefix) => {
+module.exports.run = async (client, message, args) => {
     const user = message.mentions.users.first() || message.author;
+
+    const avatar = `https://cdn.discordapp.com/avatars/${user.id}/${user.avatar}.png?size=1024`;
 
     const embed = new Discord.MessageEmbed()
         .setColor("RANDOM")
-        .setTitle(`Аватарка ${user.tag}`)
-        .setImage(`https://cdn.discordapp.com/avatars/${user.id}/${user.avatar}.png?size=1024`)
+        .setTitle(`Гей ${user.tag}`)
+        .setImage(`https://babu-api.herokuapp.com/api/gay?image=${avatar}`)
         .setTimestamp()
         .setFooter({ text: "red-lamp bot", iconURL: "https://i.imgur.com/0PCQtit.png" });
     message.channel.send({ embeds: [embed] });
 };
 
 module.exports.help = {
-    name: "avatar",
-    aliases: ["ava"],
+    name: "gay",
+    // aliases: [""]
 };