+++ mongo.js
@@ -1,8 +1,8 @@
-const mongoose = require('mongoose')
-require('dotenv').config();
-const mongoPath = process.env.MONGODB_URI
+const mongoose = require("mongoose");
+require("dotenv").config();
+const mongoPath = process.env.MONGODB_URI;
 
 module.exports = async () => {
-    await mongoose.connect(mongoPath)
-    return mongoose
-}
+    await mongoose.connect(mongoPath);
+    return mongoose;
+};